+++ src/main/scala/org/mmisw/orr/ont/swld/ontUtil.scala
@@ -12,6 +12,7 @@ import org.json4s._
 import org.mmisw.orr.ont.vocabulary.{Omv, OmvMmi}
 
 import scala.collection.JavaConversions._
+import scala.util.{Failure, Success, Try}
 
 
 object ontUtil extends AnyRef with Logging {
@@ -23,7 +24,8 @@ object ontUtil extends AnyRef with Logging {
       "rdf"     -> "application/rdf+xml"    // https://www.w3.org/TR/REC-rdf-syntax/
     , "owl"     -> "application/rdf+xml"    // https://www.w3.org/TR/REC-rdf-syntax/
     , "owx"     -> "application/owl+xml"    // https://www.w3.org/TR/owl-xml-serialization/
-    , "v2r"     -> "application/v2r+json"
+    , "v2r"     -> "application/v2r+json"   // custom ORR vocabulary format
+    , "m2r"     -> "application/m2r+json"   // custom ORR mapping format
     , "jsonld"  -> "application/json+ld"    // http://www.ietf.org/rfc/rfc6839.txt
     , "n3"      -> "text/n3"                // http://www.w3.org/TeamSubmission/n3/
     , "ttl"     -> "text/turtle"            // http://www.w3.org/TeamSubmission/turtle/
@@ -52,12 +54,13 @@ object ontUtil extends AnyRef with Logging {
   }
 
   // TODO review along with mapping in storedFormat method
-  val storedFormats = List("rdf", "n3", "owx", "jsonld", "v2r")
+  val storedFormats = List("rdf", "n3", "owx", "jsonld", "v2r", "m2r")
 
   // for the files actually stored
   def storedFormat(format: String) = format.toLowerCase match {
     case "owx"              => "owx"    // https://www.w3.org/TR/owl-xml-serialization/
     case "v2r"              => "v2r"
+    case "m2r"              => "m2r"
     case "owl"  | "rdf"     => "rdf"
     case "json" | "jsonld"  => "jsonld"
     case "ttl"  | "n3"      => "n3"
@@ -126,13 +129,13 @@ object ontUtil extends AnyRef with Logging {
         val propUri = prop.getURI
         val node: RDFNode = stmt.getObject
         val nodeString = nodeAsString(node)
-
-        val newValues = nodeString :: (map.get(propUri) match {
-          case None         => List()
-          case Some(values) => values
-        })
-
-        map = map.updated(propUri, newValues)
+        if (nodeString != null) {
+          val newValues = nodeString :: (map.get(propUri) match {
+            case None         => List()
+            case Some(values) => values
+          })
+          map = map.updated(propUri, newValues)
+        }
       }
     }
     map
@@ -225,6 +228,9 @@ object ontUtil extends AnyRef with Logging {
     else if ("V2R" == lang) {
       v2r.loadOntModel(file, Some(uri)).ontModel
     }
+    else if ("M2R" == lang) {
+      m2r.loadOntModel(file, Some(uri)).ontModel
+    }
     else {
       val ontModel = createDefaultOntModel
       ontModel.setDynamicImports(false)
@@ -252,6 +258,7 @@ object ontUtil extends AnyRef with Logging {
   def format2lang(format: String) = Option(format.toLowerCase match {
     case "owx"          => "OWX"      // to use OWL API
     case "v2r"          => "V2R"      // see v2r module
+    case "m2r"          => "M2R"      // see m2r module
     case "owl" | "rdf"  => "RDF/XML"
     case "jsonld"       => "JSON-LD"
     case "n3"           => "N3"
@@ -439,4 +446,35 @@ object ontUtil extends AnyRef with Logging {
 
     (array orElse primitive)(jValue)
   }
+
+  def getOntologySubjects(ontModel: OntModel, excludeUri: String): Map[String, Map[String, AnyRef]] = {
+    val map = scala.collection.mutable.HashMap[String, Map[String, AnyRef]]()
+    val it = ontModel.listSubjects()
+    if (it != null) while (it.hasNext) {
+      val resource = it.nextResource()
+      val resourceUri = resource.getURI
+      if (!resource.isAnon && excludeUri != resourceUri) {
+        map += (resourceUri -> extractAttributes(resource))
+      }
+    }
+    map.toMap
+  }
+
+  // TODO make this operation async (a general TODO actually)
+  def loadExternalModel(uri: String): Try[OntModel] = {
+    logger.debug(s"loadExternalModel: uri=$uri")
+
+    val ontModel = createDefaultOntModel
+    ontModel.setDynamicImports(false)
+    ontModel.getDocumentManager.setProcessImports(false)
+    try {
+      ontModel.read(uri)
+      logger.debug(s"loadExternalModel: uri=$uri: ontModel read complete.")
+      Success(ontModel)
+    }
+    catch {
+      case t: Throwable => Failure(t)
+    }
+  }
+
 }