+++ src/test/scala/org/mmisw/orr/ont/app/SequenceSpec.scala
@@ -528,7 +528,6 @@ class SequenceSpec extends MutableScalatraSpec with BaseSpec with Mockito with L
       }
     }
 
-    // TODO "succeed with no explicit org"?
   }
 
   "Register a new ont (POST /ont) with embedded ontology contents" should {
@@ -553,7 +552,40 @@ class SequenceSpec extends MutableScalatraSpec with BaseSpec with Mockito with L
         res.uri must_== embeddedUri
       }
     }
+  }
 
+  "Register a new v2r ont (POST /ont)" should {
+    "succeed" in {
+      // need a diff uri
+      val v2rMap = Map("uri" -> newOntUri(),
+        "name" -> "a v2r ontology",
+        "userName" -> userName,
+        "format" -> "v2r"
+      )
+      val v2rFile = new File("src/test/resources/vr1.v2r")
+      post("/ont", v2rMap, Map("file" -> v2rFile), headers = adminHeaders) {
+        status must_== 201
+        val res = parse(body).extract[OntologyRegistrationResult]
+        res.uri must_== v2rMap("uri")
+      }
+    }
+  }
+
+  "Register a new m2r ont (POST /ont)" should {
+    "succeed" in {
+      // need a diff uri
+      val m2rMap = Map("uri" -> newOntUri(),
+        "name" -> "a m2r ontology",
+        "userName" -> userName,
+        "format" -> "m2r"
+      )
+      val m2rFile = new File("src/test/resources/mr1.m2r")
+      post("/ont", m2rMap, Map("file" -> m2rFile), headers = adminHeaders) {
+        status must_== 201
+        val res = parse(body).extract[OntologyRegistrationResult]
+        res.uri must_== m2rMap("uri")
+      }
+    }
   }
 
   "Get an ont with given uri (GET /ont)" should {